requirements.txt
playwright==1.40.0
selenium==4.15.2
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.7
pandas==2.1.3
undetected-chromedriver==3.5.4
fake-useragent==1.4.0
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
lxml==4.9.3
selenium-stealth==1.0.6
webdriver-manager==4.0.1 